Crate fedimint_client_legacy

Modules

api

db

ln

mint

modules

outcome

serde_keypair

transaction

utils

wallet

Structs

Client

ClientSecret

GatewayClientConfig

PaymentParameters

UserClientConfig

Enums

ClientError

ConfigVerifyError

Constants

MINT_SECRET_CHILD_ID

Mint module’s secret key derivation child id

OUTGOING_LN_CONTRACT_TIMELOCK 🔒

Number of blocks until outgoing lightning contracts times out and user client can get refund

Functions

module_decode_stubs

Builds a fake module registry which is only usable for decoding messages since the client isn’t modularized yet but we need the decoding functionality.

Type Definitions

GatewayClient

Result

UserClient