Module fedimint_client_legacy::mint

Modules

backup

Backup and recovery of ecash notes

db

Structs

MintClient

Federation module client for the Mint module. It can both create transaction inputs and outputs of the mint type.

NoteIndex

An index used to deterministically derive [Note]s

NoteIssuanceRequest

Single [Note] issuance request to the mint.f

NoteIssuanceRequests

Multiple [Note] issuance requests

SpendableNote

A [Note] with associated secret key that allows to proof ownership (spend it)

Enums

MintClientError

NoteFinalizationError

Constants

MINT_E_CASH_BACKUP_SNAPSHOT_TYPE_CHILD_ID 🔒

MINT_E_CASH_FETCH_TIMEOUT 🔒

MINT_E_CASH_TYPE_CHILD_ID 🔒

Type Definitions

Result 🔒