Struct fedimint_client_legacy::Client

pub struct Client<C> {
    pub(crate) config: C,
    pub(crate) context: Arc<ClientContext>,
    pub(crate) root_secret: DerivableSecret,
}

Fields
config: Ccontext: Arc<ClientContext>root_secret: DerivableSecret
Implementations

impl<C> Client<C>

pub fn decoders(&self) -> &ModuleDecoderRegistry

pub fn module_gens(&self) -> &ClientModuleGenRegistry

pub fn context(&self) -> &Arc<ClientContext>

pub fn root_secret(&self) -> &DerivableSecret

impl<T> Client<T>

pub fn mint_secret_static(root_secret: &DerivableSecret) -> DerivableSecret

impl<T: AsRef<ClientConfig> + Clone + Send> Client<T>

pub fn db(&self) -> &Database

pub fn ln_client(&self) -> LnClient

pub fn mint_client(&self) -> MintClient

pub fn wallet_client(&self) -> WalletClient

pub fn config(&self) -> T

pub async fn new(
config: T,
decoders: ModuleDecoderRegistry,
module_gens: ClientModuleGenRegistry,
db: Database,
secp: Secp256k1<All>
) -> Self

pub async fn new_with_api(
config: T,
decoders: ModuleDecoderRegistry,
module_gens: ClientModuleGenRegistry,
db: Database,
api: DynGlobalApi,
secp: Secp256k1<All>
) -> Client<T>

pub(crate) async fn get_secret(db: &Database) -> DerivableSecret

Fetches the client secret from the database or generates a new one if none is present

pub async fn peg_in<R: RngCore + CryptoRng>(
&self,
txout_proof: TxOutProof,
btc_transaction: BitcoinTransaction,
rng: R
) -> Result<TransactionId>

pub async fn submit_tx_with_change<R: RngCore + CryptoRng>(
&self,
tx: TransactionBuilder,
rng: R
) -> Result<TransactionId>

Submits a transaction to the fed, making change using our change module

TODO: For safety, if the submission fails, the DB write still occurs. We should instead ensure the state of the client and consensus are always the same.

pub async fn reissue<R: RngCore + CryptoRng>(
&self,
notes: TieredMulti<SpendableNote>,
rng: R
) -> Result<OutPoint>

Spent some SpendableNotes to receive a freshly minted ones

This is useful in scenarios where certain notes were handed over directly to us by another user as a payment. By spending them we can make sure they can no longer be potentially double-spent.

On success the out point of the newly issued e-cash notes is returned. It can be used to easily poll the transaction status using MintClient::fetch_notes until it returns Ok(()), indicating we received our newly issued e-cash notes.

pub async fn validate_note_signatures(
&self,
notes: &TieredMulti<SpendableNote>
) -> Result<()>

Validate signatures on notes.

This function checks if signatures are valid based on the federation public key. It does not check if the nonce is unspent.

pub async fn pay_to_blind_nonces<R: RngCore + CryptoRng>(
&self,
blind_nonces: TieredMulti<BlindNonce>,
rng: R
) -> Result<OutPoint>

Pay by creating notes provided (and most probably controlled) by the recipient.

A standard way to facilitate a payment between users of a mint. Generate a transaction spending notes we own as inputs and creating new notes from [BlindNonce]s provided by the recipient as outputs.

Returns a OutPoint of a fedimint transaction created and submitted as a payment.

The name is derived from Bitcoin’s terminology of “pay to <address-type>”.

pub async fn receive_notes(
&self,
amount: Amount
) -> (TieredMulti<BlindNonce>, Box<dyn Fn(OutPoint)>)

Receive e-cash directly from another user when online (vs. offline transfer)

Generates notes that another user will pay for and let us know the OutPoint in create_tx Payer can use the pay_to_blind_nonces function Allows transfer of e-cash without risk of double-spend or not having exact change

pub async fn new_peg_out_with_fees(
&self,
amount: Amount,
recipient: Address
) -> Result<PegOut>

pub async fn rbf_tx(&self, rbf: Rbf) -> Result<OutPoint>

pub async fn peg_out<R: RngCore + CryptoRng>(
&self,
peg_out: PegOut,
rng: R
) -> Result<OutPoint>

pub async fn get_new_pegin_address<R: RngCore + CryptoRng>(
&self,
rng: R
) -> Address

Returns a bitcoin address suited to perform a fedimint peg-in

This function requires a cryptographically secure randomness source, and utilizes the wallet-clients get_new_pegin_address to derive a bitcoin-address from the federations public descriptor by tweaking it.

this function will write to the clients DB

read more on fedimints address derivation: https://fedimint.org/Fedimint/wallet/

pub async fn spend_ecash<R: RngCore + CryptoRng>(
&self,
amount: Amount,
rng: R
) -> Result<TieredMulti<SpendableNote>>

Issues a spendable amount of ecash

WARNING the ecash will be deleted from the database, the returned ecash must be reissued or it will be lost

pub async fn remove_ecash(&self, ecash: TieredMulti<SpendableNote>)

Removes spent ecash from the database

pub async fn remint_ecash<R: RngCore + CryptoRng>(
&self,
amount: Amount,
rng: R
) -> Result<()>

For tests only: Select notes of a given amount, and then remint them, remove the amount of notes from the database and return it to the user.

This is a respin of spent_ecash for tests, where it is necessary to process epochs after self.submit_tx_with_change. Then remint_ecash_await can be called to do the rest.

TODO: Like spend_ecash, I think this function works in tests mostly by accident.

pub async fn remint_ecash_await(
&self,
amount: Amount
) -> Result<TieredMulti<SpendableNote>>

Continuation of remint_notes

pub async fn fetch_notes<'a>(&self, outpoint: OutPoint) -> Result<()>

Tries to fetch e-cash notes from a certain out point. An error may just mean having queried the federation too early. Use MintClientError::is_retryable to determine if the operation should be retried at a later time.

pub async fn reissue_pending_notes<R: RngCore + CryptoRng>(
&self,
rng: R
) -> Result<OutPoint>

Should be called after any transaction that might have failed in order to get any note inputs back.

pub async fn await_consensus_block_height(
&self,
block_height: u64
) -> Result<u64, Elapsed>

pub(crate) async fn await_consensus_block_height_inner(
&self,
block_height: u64
) -> u64

pub async fn fetch_all_notes<'a>(&self) -> Result<Vec<OutPoint>>

pub async fn summary(&self) -> TieredSummary

pub async fn notes(&self) -> TieredMulti<SpendableNote>

pub async fn list_active_issuances(
&self
) -> Vec<(OutPoint, NoteIssuanceRequests)>

pub async fn fetch_epoch_history(
&self,
epoch: u64,
epoch_pk: PublicKey
) -> Result<SignedEpochOutcome>

impl Client<UserClientConfig>

pub async fn fetch_registered_gateways(&self) -> Result<Vec<LightningGateway>>

pub async fn fetch_active_gateway(&self) -> Result<LightningGateway>

pub async fn switch_active_gateway(
&self,
node_pub_key: Option<PublicKey>
) -> Result<LightningGateway>

Switches the clients active gateway to a registered gateway with the given node pubkey. If no pubkey is given (node_pub_key == None) the first available registered gateway is activated. This behavior is useful for scenarios where we don’t know any registered gateways in advance.

pub async fn fund_outgoing_ln_contract<R: RngCore + CryptoRng>(
&self,
invoice: Invoice,
rng: R
) -> Result<(ContractId, OutPoint)>

pub async fn try_refund_outgoing_contract(
&self,
contract_id: ContractId,
rng: impl RngCore + CryptoRng
) -> Result<OutPoint>

Claims a refund for an expired or cancelled outgoing contract

This can be necessary when the Lightning gateway cannot route the payment, is malicious or offline. The function returns the out point of the e-cash output generated as change.

pub async fn await_outgoing_contract_acceptance(
&self,
outpoint: OutPoint
) -> Result<()>

pub async fn await_outpoint_outcome(&self, outpoint: OutPoint) -> Result<()>

Waits for the federation to sign an ecash note.

This function will poll until the returned result includes a SigResponse from the federation or it will timeout.

pub async fn generate_confirmed_invoice<R: RngCore + CryptoRng>(
&self,
amount: Amount,
description: String,
rng: R,
expiry_time: Option<u64>
) -> Result<ConfirmedInvoice>

pub async fn generate_unconfirmed_invoice_and_submit<R: RngCore + CryptoRng>(
&self,
amount: Amount,
description: String,
rng: R,
expiry_time: Option<u64>
) -> Result<(TransactionId, Invoice, KeyPair)>

pub async fn generate_unconfirmed_invoice<R: RngCore + CryptoRng>(
&self,
amount: Amount,
description: String,
payment_keypair: KeyPair,
rng: R,
expiry_time: Option<u64>
) -> Result<(Invoice, Output)>

pub async fn await_invoice_confirmation(
&self,
txid: TransactionId,
invoice: Invoice,
payment_keypair: KeyPair
) -> Result<ConfirmedInvoice>

pub async fn claim_incoming_contract(
&self,
contract_id: ContractId,
rng: impl RngCore + CryptoRng
) -> Result<OutPoint>

pub async fn await_outgoing_contract_execution(
&self,
contract_id: ContractId,
rng: impl RngCore + CryptoRng
) -> Result<()>

Notify gateway that we’ve escrowed notes they can claim by routing our payment and wait for them to do so

impl Client<GatewayClientConfig>

pub async fn fetch_outgoing_contract(
&self,
contract_id: ContractId
) -> Result<OutgoingContractAccount>

Fetch the specified outgoing payment contract account

pub async fn validate_outgoing_account(
&self,
account: &OutgoingContractAccount
) -> Result<PaymentParameters>

Check if we can claim the contract account and returns the max delay in blocks for how long other nodes on the route are allowed to delay the payment.

pub(crate) fn is_maybe_internal_payment(&self, invoice: &Invoice) -> bool

Returns true if the invoice contains us as a routing hint

pub async fn save_outgoing_payment(&self, contract: OutgoingContractAccount)

Save the details about an outgoing payment the client is about to process. This function has to be called prior to instructing the lightning node to pay the invoice since otherwise a crash could lead to loss of funds.

Note though that extended periods of staying offline will result in loss of funds anyway if the client can not claim the respective contract in time.

pub async fn list_pending_outgoing(&self) -> Vec<OutgoingContractAccount>

Lists all previously saved transactions that have not been driven to completion so far

pub async fn abort_outgoing_payment(&self, contract_id: ContractId) -> Result<()>

Abort payment if our node can’t route it and give money back to user

pub async fn cancel_outgoing_contract(
&self,
contract_account: OutgoingContractAccount
) -> Result<()>

Cancel an outgoing contract we haven’t accepted yet, possibly because it was underfunded

pub async fn claim_outgoing_contract(
&self,
contract_id: ContractId,
preimage: Preimage,
rng: impl RngCore + CryptoRng
) -> Result<OutPoint>

Claim an outgoing contract after acquiring the preimage by paying the associated invoice and initiates e-cash issuances to receive the bitcoin from the contract (these still need to be fetched later to finalize them).

Callers need to make sure that the contract can still be claimed by the gateway and has not timed out yet. Otherwise the transaction will fail.

pub async fn buy_preimage_offer(
&self,
payment_hash: &Hash,
htlc_amount: &Amount,
rng: impl RngCore + CryptoRng
) -> Result<(OutPoint, ContractId)>

Buy a lightning preimage listed for sale inside the federation

Called when a lightning gateway attempts to satisfy a contract on behalf of a user

payment_hash - hash of the preimage we want to buy. It is included inside a bolt11 invoice and should match the offer hash

htlc_amount - amount from the htlc the gateway wants to pay. Should be less than or equal to the offer amount depending on gateway fee policy

pub async fn refund_incoming_contract(
&self,
contract_id: ContractId,
rng: impl RngCore + CryptoRng
) -> Result<TransactionId>

Claw back funds after incoming contract that had invalid preimage

pub async fn list_pending_claimed_outgoing(&self) -> Vec<ContractId>

Lists all claim transactions for outgoing contracts that we have submitted but were not part of the consensus yet.

pub async fn await_preimage_decryption(
&self,
outpoint: OutPoint
) -> Result<Preimage>

Wait for a lightning preimage gateway has purchased to be decrypted by the federation

pub async fn await_outgoing_contract_claimed(
&self,
contract_id: ContractId,
outpoint: OutPoint
) -> Result<()>

Waits for a outgoing contract claim transaction to be confirmed and retransmits it periodically if this does not happen.

pub async fn list_fetchable_notes(&self) -> Vec<OutPoint>

pub async fn register_with_federation(
&self,
config: LightningGateway
) -> Result<()>

Register this gateway with the federation