Type Alias fedimint_client_legacy::UserClient

pub type UserClient = Client<UserClientConfig>;

Aliased Type

struct UserClient {
    pub(crate) concurrency_lock: Mutex<ConcurrencyLock>,
    pub(crate) config: UserClientConfig,
    pub(crate) context: Arc<ClientContext, Global>,
    pub(crate) root_secret: DerivableSecret,
}

Fields
concurrency_lock: Mutex<ConcurrencyLock>config: UserClientConfigcontext: Arc<ClientContext, Global>root_secret: DerivableSecret
Implementations

impl<C> Client<C>

pub async fn concurrency_lock(&self) -> MutexGuard<'_, ConcurrencyLock>

pub fn decoders(&self) -> &ModuleDecoderRegistry

pub fn module_inits(&self) -> &ClientModuleInitRegistry

pub fn context(&self) -> &Arc<ClientContext>

pub fn root_secret(&self) -> &DerivableSecret

impl<T> Client<T>

pub fn mint_secret_static(root_secret: &DerivableSecret) -> DerivableSecret

impl<T: AsRef<ClientConfig> + Clone + MaybeSend> Client<T>

pub fn db(&self) -> &Database

pub fn ln_client(&self) -> LnClient

pub fn mint_client(&self) -> MintClient

pub fn wallet_client(&self) -> WalletClient

pub fn config(&self) -> T

pub fn config_ref(&self) -> &T

pub async fn new( config: T, decoders: ModuleDecoderRegistry, module_inits: ClientModuleInitRegistry, db: Database, secp: Secp256k1<All> ) -> Self

pub async fn new_with_api( config: T, decoders: ModuleDecoderRegistry, module_inits: ClientModuleInitRegistry, db: Database, api: DynGlobalApi, secp: Secp256k1<All> ) -> Client<T>

pub(crate) async fn get_secret(db: &Database) -> DerivableSecret

Fetches the client secret from the database or generates a new one if none is present

pub async fn peg_in<R: RngCore + CryptoRng>( &self, txout_proof: TxOutProof, btc_transaction: BitcoinTransaction, rng: R ) -> Result<TransactionId>

pub async fn submit_tx_with_change<R: RngCore + CryptoRng>( &self, guard: MutexGuard<'_, ConcurrencyLock>, tx: TransactionBuilder, rng: R ) -> Result<TransactionId>

Submits a transaction to the fed, making change using our change module

TODO: For safety, if the submission fails, the DB write still occurs. We should instead ensure the state of the client and consensus are always the same.

pub async fn reissue<R: RngCore + CryptoRng>( &self, notes: TieredMulti<SpendableNote>, rng: R ) -> Result<OutPoint>

Spent some SpendableNotes to receive a freshly minted ones

This is useful in scenarios where certain notes were handed over directly to us by another user as a payment. By spending them we can make sure they can no longer be potentially double-spent.

On success the out point of the newly issued e-cash notes is returned. It can be used to easily poll the transaction status using MintClient::fetch_notes until it returns Ok(()), indicating we received our newly issued e-cash notes.

pub async fn validate_note_signatures( &self, notes: &TieredMulti<SpendableNote> ) -> Result<()>

Validate signatures on notes.

This function checks if signatures are valid based on the federation public key. It does not check if the nonce is unspent.

pub async fn pay_to_blind_nonces<R: RngCore + CryptoRng>( &self, blind_nonces: TieredMulti<BlindNonce>, rng: R ) -> Result<OutPoint>

Pay by creating notes provided (and most probably controlled) by the recipient.

A standard way to facilitate a payment between users of a mint. Generate a transaction spending notes we own as inputs and creating new notes from [BlindNonce]s provided by the recipient as outputs.

Returns a OutPoint of a fedimint transaction created and submitted as a payment.

The name is derived from Bitcoin’s terminology of “pay to <address-type>”.

pub async fn receive_notes( &self, amount: Amount ) -> (TieredMulti<BlindNonce>, Box<dyn Fn(OutPoint)>)

Receive e-cash directly from another user when online (vs. offline transfer)

Generates notes that another user will pay for and let us know the OutPoint in create_tx Payer can use the pay_to_blind_nonces function Allows transfer of e-cash without risk of double-spend or not having exact change

pub async fn new_peg_out_with_fees( &self, amount: Amount, recipient: Address ) -> Result<PegOut>

pub async fn rbf_tx(&self, rbf: Rbf) -> Result<OutPoint>

pub async fn peg_out<R: RngCore + CryptoRng>( &self, peg_out: PegOut, rng: R ) -> Result<OutPoint>

pub async fn get_new_pegin_address<R: RngCore + CryptoRng>( &self, rng: R ) -> Address

Returns a bitcoin address suited to perform a fedimint peg-in

This function requires a cryptographically secure randomness source, and utilizes the wallet-clients get_new_pegin_address to derive a bitcoin-address from the federations public descriptor by tweaking it.

this function will write to the clients DB

read more on fedimints address derivation: https://fedimint.org/Fedimint/wallet/

pub async fn spend_ecash<R: RngCore + CryptoRng>( &self, amount: Amount, rng: R ) -> Result<TieredMulti<SpendableNote>>

Issues a spendable amount of ecash

WARNING the ecash will be deleted from the database, the returned ecash must be reissued or it will be lost

pub async fn remove_ecash(&self, ecash: TieredMulti<SpendableNote>)

Removes spent ecash from the database

pub async fn remint_ecash<R: RngCore + CryptoRng>( &self, amount: Amount, rng: R ) -> Result<()>

For tests only: Select notes of a given amount, and then remint them, remove the amount of notes from the database and return it to the user.

This is a respin of spent_ecash for tests, where it is necessary to process epochs after self.submit_tx_with_change. Then remint_ecash_await can be called to do the rest.

TODO: Like spend_ecash, I think this function works in tests mostly by accident.

pub async fn remint_ecash_await( &self, amount: Amount ) -> Result<TieredMulti<SpendableNote>>

Continuation of remint_notes

pub async fn fetch_notes<'a>(&self, outpoint: OutPoint) -> Result<()>

Tries to fetch e-cash notes from a certain out point. An error may just mean having queried the federation too early. Use MintClientError::is_retryable to determine if the operation should be retried at a later time.

pub async fn reissue_pending_notes<R: RngCore + CryptoRng>( &self, rng: R ) -> Result<OutPoint>

Should be called after any transaction that might have failed in order to get any note inputs back.

pub async fn await_consensus_block_count( &self, block_count: u64 ) -> Result<u64, Elapsed>

pub(crate) async fn await_consensus_block_count_inner( &self, requested_count: u64 ) -> u64

pub async fn fetch_all_notes<'a>(&self) -> Result<Vec<OutPoint>>

pub async fn summary(&self) -> TieredSummary

pub async fn notes(&self) -> TieredMulti<SpendableNote>

pub async fn list_active_issuances( &self ) -> Vec<(OutPoint, NoteIssuanceRequests)>

pub async fn fetch_epoch_history( &self, epoch: u64, epoch_pk: PublicKey ) -> Result<SignedEpochOutcome>

impl Client<UserClientConfig>

pub async fn fetch_registered_gateways(&self) -> Result<Vec<LightningGateway>>

pub async fn fetch_active_gateway(&self) -> Result<LightningGateway>

pub async fn switch_active_gateway( &self, gateway_id: Option<PublicKey> ) -> Result<LightningGateway>

Switches the clients active gateway to a registered gateway with the given node pubkey. If no pubkey is given (node_pub_key == None) the first available registered gateway is activated. This behavior is useful for scenarios where we don’t know any registered gateways in advance.

pub async fn fund_outgoing_ln_contract<R: RngCore + CryptoRng>( &self, invoice: Bolt11Invoice, rng: R ) -> Result<(ContractId, OutPoint)>

pub async fn try_refund_outgoing_contract( &self, contract_id: ContractId, rng: impl RngCore + CryptoRng ) -> Result<OutPoint>

Claims a refund for an expired or cancelled outgoing contract

This can be necessary when the Lightning gateway cannot route the payment, is malicious or offline. The function returns the out point of the e-cash output generated as change.

pub async fn await_outgoing_contract_acceptance( &self, outpoint: OutPoint ) -> Result<()>

pub async fn await_outpoint_outcome(&self, outpoint: OutPoint) -> Result<()>

Waits for the federation to sign an ecash note.

This function will poll until the returned result includes a SigResponse from the federation or it will timeout.

pub async fn generate_confirmed_invoice<R: RngCore + CryptoRng>( &self, amount: Amount, description: String, rng: R, expiry_time: Option<u64> ) -> Result<ConfirmedInvoice>

pub async fn generate_unconfirmed_invoice_and_submit<R: RngCore + CryptoRng>( &self, amount: Amount, description: String, rng: R, expiry_time: Option<u64> ) -> Result<(TransactionId, Bolt11Invoice, KeyPair)>

pub async fn generate_unconfirmed_invoice<R: RngCore + CryptoRng>( &self, amount: Amount, description: String, payment_keypair: KeyPair, rng: R, expiry_time: Option<u64> ) -> Result<(Bolt11Invoice, Output)>

pub async fn await_invoice_confirmation( &self, txid: TransactionId, invoice: Bolt11Invoice, payment_keypair: KeyPair ) -> Result<ConfirmedInvoice>

pub async fn claim_incoming_contract( &self, contract_id: ContractId, rng: impl RngCore + CryptoRng ) -> Result<OutPoint>

pub async fn await_outgoing_contract_execution( &self, contract_id: ContractId, rng: impl RngCore + CryptoRng ) -> Result<()>

Notify gateway that we’ve escrowed notes they can claim by routing our payment and wait for them to do so